2008-06-08  Marco van den Heuvel  <blackystardust68@yahoo.com>
	* configure.in: Added a check for the minix version for the
	prototypes of recv and send in network.c.

	* doc/html/solaris.html, doc/html/amigaos.html,
	doc/html/sco.html, doc/html/hpux.html, doc/html/beos.html,
	doc/html/qnx.html, doc/html/Makefile.am, doc/html/index.html:
	Updated for the upcoming 2.0 release.

	* po/nl.po, po/POTFILES.in: Updated.

	* src/gfxoutputdrv/ffmpegdrv.c: Corrected a typo.

	* src/vdrive/vdrive-bam.h, src/vdrive/vdrive-bam.c: Corrected
	block allocation and interleave for c1541/vdrive as provided by
	Ingo Korb <ingo@akana.de>.

	* src/video/render2x2pal.c, src/video/render2x2.c,
	src/video/render2x2pal.h, src/video/video-resources.c,
	src/video/render2x2.h, src/video/video-color.c,
	src/video/video-resources.h, src/video/video-resources-pal.c,
	src/video/video-render-pal.c, src/video.h,
	src/vicii/vicii-cmdline-options.c: Added new PAL emulation as
	provided by Groepaz/Hitmen <groepaz@gmx.net>.

	* src/network.c: Made the recv and send prototypes for minix
	only be used for minix versions that don't have the prototypes
	in the system headers.

	* src/sounddrv/soundbeos.cc, src/initcmdline.c,
	src/arch/beos/archdep.c, src/arch/beos/ui_ram.cc,
	src/arch/beos/ui_sound.cc, src/arch/beos/ui_video.cc,
	src/arch/beos/kbd.c, src/arch/beos/ui_vicii.cc,
	src/arch/beos/joy.cc, src/arch/beos/kbd.h,
	src/arch/beos/ui_sid.cc, src/arch/beos/vicewindow.cc,
	src/arch/beos/constants.h, src/arch/beos/ui_device.cc,
	src/arch/beos/vicemenu.cc, src/arch/beos/uimon.c,
	src/arch/beos/blockdev.c, src/arch/beos/c64ui.cc: Fixed some
	beos warnings/issues.

	* src/c64/c64io.c, src/c64/c64io.h,
	src/c64/cart/actionreplay4.c, src/c64/cart/actionreplay4.h,
	src/c64/cart/crt.c, src/c64/cart/c64cartmem.c,
	src/c64/cart/c64cart.c, src/c64/cart/stardos.c,
	src/c64/cart/stardos.h, src/c64/cart/Makefile.am,
	src/cartridge.h: Added the action replay 4 and stardos carts, as
	provided by Groepaz/Hitmen <groepaz@gmx.net>.

	* src/c64/c64-resources.c, src/c64/c64rom.c, src/c64/c64rom.h,
	src/c128/c128rom.c, src/c128/c128rom.h,
	src/c128/c128-resources.c: Added a parameter to the
	c64rom_load_kernal function to be able to have hardware kernal
	replacements (without using ultimax), as used in the stardos
	cart.

	* src/translate.txt, src/translate.c, src/arch/amigaos/intl.c,
	src/arch/amigaos/intl.txt, src/arch/win32/intl.c: Added
	translation elements for the new PAL emulation and new carts.

	* src/sound.c: Added a fix to not limit the minimum size of the
	sound buffer as provided by Hannu Nuotio <hannu.nuotio@tut.fi>.

	* src/arch/amigaos/mui/uivideoc128.c,
	src/arch/amigaos/mui/uivideoc64plus4vic20.c,
	src/arch/amigaos/c64uires.h, src/arch/amigaos/c128uires.h,
	src/arch/amigaos/uic64cart.c, src/arch/amigaos/uires.h,
	src/arch/win32/resmenu.txt, src/arch/win32/uic64cart.c,
	src/arch/win32/resstrings.txt, src/arch/win32/resdialogs.txt,
	src/arch/win32/resc128.rc, src/arch/win32/resc64.rc,
	src/arch/win32/uivideo.c, src/arch/win32/res.rc,
	src/arch/os2/dialogs/dlg-color.rc,
	src/arch/os2/dialogs/menubar.rc, src/arch/os2/dialogs/menubar.c,
	src/arch/os2/dialogs/menubar.h,
	src/arch/os2/dialogs/dlg-color.c,
	src/arch/os2/dialogs/dlg-color.h, src/arch/unix/gui/uipalemu.c,
	src/arch/unix/gui/uic64cart.c,
	src/arch/unix/x11/gnome/uipalcontrol.c,
	src/arch/msdos/uic64cart.c: Added UI elements for the new PAL
	emulation and the new carts.

	* src/arch/amigaos/make-bindist.sh,
	src/arch/unix/sco_sol/make-bindist.sh,
	src/arch/unix/gp2x/make-bindist.sh,
	src/arch/unix/qnx6/make-bindist.sh,
	src/arch/msdos/make-bindist.sh, src/arch/beos/make-bindist.sh:
	Fixed to not include the new win32 shortcut file in the binary
	dist.

	* src/arch/win32/watcom/c64cart.tgt, src/arch/win32/c64cart.dsp,
	src/arch/os2/watcom/c64cart.tgt,
	src/arch/os2/vac++/incl/c64cart.icc: Added the new cart files to
	the project files.

	* src/arch/unix/vsyncarch.c, src/arch/unix/rs232.c,
	src/arch/unix/coproc.c, src/arch/unix/signals.c: Fixed some
	openserver 6.x issues.

	* src/arch/unix/gui/c64ui.c, src/arch/unix/gui/uivicii.c,
	src/arch/unix/gui/uivicii.h, src/arch/unix/gui/c128ui.c,
	src/arch/unix/gui/cbm2ui.c: Fixed the set_video_standard_submenu
	issue.

	* src/arch/unix/mousedrv.c, src/arch/unix/x11/xaw/x11ui.c: Added
	xaw mouse grab fix as provided by Hannu Nuotio
	<hannu.nuotio@tut.fi>.

	* src/arch/unix/x11/qnxshm.c: Fixed some qnx warnings/issues.

	* src/arch/unix/x11/xaw/widgets/FileSel.c,
	src/arch/unix/x11/xaw/widgets/MultiList.c: Added min/max
	definition fix.

	* src/arch/unix/x11/renderxv.c: Fixed some minix
	warnings/issues.

	* src/arch/beos/console.c: Added beos console fix as provided by
	Marcus Sutton <loggedoubt@gmail.com>

	* src/cartconv.c: Added support for the new carts.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@18859 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/beos/make-bindist.sh
@@ -56,7 +56,11 @@ rm `find BeVICE-$VICEVERSION.$BEOSCPU -name "os2*.vkm"`
 rm `find BeVICE-$VICEVERSION.$BEOSCPU -name "osx*.vkm"`
 rm `find BeVICE-$VICEVERSION.$BEOSCPU -name "win_*.vkm"`
 rm `find BeVICE-$VICEVERSION.$BEOSCPU -name "x11_*.vkm"`
-rm BeVICE-$VICEVERSION.$BEOSCPU/html/texi2html
+rm `find BeVICE-$VICEVERSION.$BEOSCPU -name "*.vsc"`
+if [ -e BeVICE-$VICEVERSION.$BEOSCPU/html/texi2html ]
+then
+  rm BeVICE-$VICEVERSION.$BEOSCPU/html/texi2html
+fi
 if test x"$ZIPKIND" = "xzip"; then
   if test x"$ZIP" = "x"; then
     zip -r -9 -q BeVICE-$VICEVERSION.$BEOSCPU.zip BeVICE-$VICEVERSION.$BEOSCPU
